% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassPerm.R
\name{ClassPerm}
\alias{ClassPerm}
\title{Permutation Analysis for classification}
\usage{
ClassPerm(Data, predictorCol, selectedCols, classifierFun, nSims = 1000, ...)
}
\details{
\enumerate{
\item Outputs
  \enumerate{
    \item p-value from permutation
    \item figure containing null distribution
}
}
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

