% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{DimensionRed}
\alias{DimensionRed}
\title{Generic Dimensionallity Reduction Function}
\usage{
DimensionRed(Data, method = "MDS", selectedCols, outcome = NA,
  plot = FALSE, ...)
}
\arguments{
\item{Data}{(dataframe) a data frame with variable/feature columns}

\item{selectedCol}{(optional)(numeric) which columns should be treated as data(features/columns) (defaults to all columns)}
}
\value{
Data frame with \code{Results}
}
\description{
A simple function to perform dimensionality reduction
}
\examples{
# reducing dimension of Grip aperture from 10 to 2
GripAperture <- DimensionRed(KinData,selectedCols = 12:21,outcome = KinData[,"Object.Size"],plot = TRUE)
}
\author{
Atesh Koul, C'MON unit, Istituto Italiano di Tecnologia

\email{atesh.koul@iit.it}
}

