% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDA.R
\name{LinearDA}
\alias{LinearDA}
\title{simple function to perform cross-validated Linear Discriminant Analysis}
\usage{
LinearDA(Data, predictorCol, selectedCols, cvType = "createDataPartition")
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

