% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCluster.R
\name{ModelCluster}
\alias{ModelCluster}
\title{Model based Clustering}
\usage{
ModelCluster(Data, NewData = NULL, G, ...)
}
\arguments{
\item{Data}{(dataframe) Data dataframe}

\item{NewData}{(optional) (dataframe) New Data frame for which the class membership is requested}

\item{G}{(optional) (numeric) No. of components to verify}
}
\value{
\code{class membership} of the clustered \code{NewData}
}
\description{
A simple function to perform Model based cluster Analysis
:
}
\examples{
# clustering kinematics data at 10\% of movement
cluster_time <- ModelCluster(KinData[,c(2,12,22,32,42,52,62,72,82,92,102,112)],G=1:12)
}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

