% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNewData.R
\name{predictNewData}
\alias{predictNewData}
\title{Predict Class membership for New Data}
\usage{
predictNewData(model, NewData, ...)
}
\arguments{
\item{model}{(model)     Classifier model obtained from a classification analysis}

\item{NewData}{(optional) (dataframe) New Data frame features for which the class membership is requested}

\item{...}{(optional) additional arguments for the function}
}
\value{
Predictions for each case in the NewData.
}
\description{
A simple function to predict class membership for new data
}
\details{
A function to generate predictions on a new dataset based on a previously estimated classifier model.
 This could be generated from LinearDA, classifyFun or DTMOdel functions.
}
\author{
Atesh Koul, C'MON unit, Istituto Italiano di Tecnologia

\email{atesh.koul@iit.it}
}

