% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredPsych-package.r
\docType{package}
\name{PredPsych}
\alias{PredPsych}
\alias{PredPsych-package}
\title{PredPsych.}
\description{
PredPsych.
}
\details{
"PredPsych" is a user-friendly, R toolbox based on machine learning predictive algorithms.
}
\author{
Atesh Koul, C'MON unit, Istituto Italiano di Tecnologia

\email{atesh.koul@iit.it}
}
\references{
Koul, A., Becchio, C., & Cavallo, A. (2017, March 21). 
PredPsych: A toolbox for predictive machine learning based approach in experimental 
psychology research. Retrieved from osf.io/preprints/psyarxiv/pvjac
}

