% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CartModel.R
\name{CartModel}
\alias{CartModel}
\title{Generic Classification and Regression Tree Function}
\usage{
CartModel(Data, responseCol, selectedCol, tree, ...)
}
\arguments{
\item{Data}{a data frame with regressors and response}

\item{responseCol}{which column should be used as response col}

\item{selectedCol}{(optional) which columns should be treated as data(features + response) (defaults to all columns)}

\item{tree}{which cart model to implement; One of the following values:
\itemize{
\item modelF  =   Full Model Tree; 
\item modelNAHF = Crossvalidated Half Model Tree removing missing values;
\item modelHF =   Crossvalidated Half Model Tree With missing values;
\item modelCF =   Conditional inference framework Tree;
\item modelRF =   Random Forest Tree;    
}}
}
\value{
Cart model result for the input tree \code{Results}
}
\description{
A simple function to create Classification and Regression Trees
}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

