% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CartModel.R
\name{CartModel}
\alias{CartModel}
\title{simple function to create Classification and Regression tree of Data}
\usage{
CartModel(Data, responseCol, selectedCol, tree, ...)
}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

