% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CartModel.R
\name{CartModel}
\alias{CartModel}
\title{Generic Classification and Regression Tree Function}
\usage{
CartModel(Data, responseCol, selectedCol, tree, ...)
}
\arguments{
\item{Data}{(dataframe) a data frame with regressors and response}

\item{responseCol}{(numeric) which column should be used as response col}

\item{selectedCol}{(optional)(numeric) which columns should be treated as data(features + response) (defaults to all columns)}

\item{tree}{which cart model to implement; One of the following values:
\itemize{
\item modelF  =   Full Model Tree; 
\item modelNAHF = Crossvalidated Half Model Tree removing missing values;
\item modelHF =   Crossvalidated Half Model Tree With missing values;
\item modelCF =   Conditional inference framework Tree;
\item modelRF =   Random Forest Tree;    
}}
}
\value{
Cart model result for the input tree \code{Results}
}
\description{
A simple function to create Classification and Regression Trees
}
\examples{
# generate a cart model for 10\% of the data with cross-validation
CaRTModel <- CartModel(Data = KinData[,c(1,2,12,22,32,42,52,62,72,82,92,102,112)],responseCol=1,tree='modelHF')



}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

