% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyFun.R
\name{classifyFun}
\alias{classifyFun}
\title{Generic Classification Analyses}
\usage{
classifyFun(Data, predictorCol, selectedCols, ranges = NULL, tune = FALSE,
  cost = 1, gamma = 0.5, classifierName = "svm",
  genclassifier = Classifier.svm, silent = FALSE, ...)
}
\arguments{
\item{Data}{(dataframe) dataframe of the data}

\item{predictorCol}{(numeric) column number that contains the variable to be predicted}

\item{selectedCols}{(optional) (numeric) all the columns of data that would be used either as predictor or as feature}

\item{ranges}{(optional) (list)  ranges for tuning support vector machine}

\item{tune}{(optional) (logical) whether tuning of svm parameters should be performed or not}

\item{cost}{(optional) (numeric) regularization parameter of svm}

\item{gamma}{(optional) (numeric)  rbf kernel parameter}

\item{classifierName}{(optional) (string) name of the classifier to be used}

\item{genclassifier}{(optional) (function or string) a classifier function or a name (e.g. Classifier.svm)}
}
\value{
Outputs Crossvalidation accuracy \code{acc} and    Test accuracy \code{accTest}
}
\description{
function for performing generic classification Analysis
}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

