% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyFun.R
\name{classifyFun}
\alias{classifyFun}
\title{simple function to create permutation testing of a classifier}
\usage{
classifyFun(Data, predictorCol, selectedCols, ranges = NULL, tune = FALSE,
  cost = 1, gamma = 0.5, classifierName = "svm",
  genclassifier = Classifier.svm, silent = FALSE, ...)
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

