% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassPerm.R
\name{ClassPerm}
\alias{ClassPerm}
\title{Permutation Analysis for classification}
\usage{
ClassPerm(Data, predictorCol, selectedCols, classifierFun, nSims = 1000, ...)
}
\arguments{
\item{Data}{(dataframe) dataframe of the data}

\item{predictorCol}{(numeric) column number that contains the variable to be predicted}

\item{selectedCols}{(optional) (numeric) all the columns of data that would be used either as predictor or as feature}

\item{classifierFun}{(optional) (function) classifier function}

\item{nSims}{(optional) (numeric) number of simulations}
}
\value{
Returns \code{actualAcc} of the classification analysis,
 \code{p-value} from permutation testing, \code{nullAcc} distribution of the permutation \code{figure} containing null distribution
}
\description{
simple function to create permutation testing of a classifier
}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

