% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassPerm.R
\name{ClassPerm}
\alias{ClassPerm}
\title{simple function to create permutation testing of a classifier}
\usage{
ClassPerm(Data, predictorCol, selectedCols, classifierFun, nSims = 1000, ...)
}
\description{
Needs the following:
:
\enumerate{
\item Inputs
  \enumerate{
    \item X = Data matrix
     \item Y = labels
     \item classifierFun = classifier function
     \item nSims = number of simulations
}
}
 \enumerate{
\item Outputs
  \enumerate{
    \item p-value from permutation
    \item figure containing null distribution
}
}
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

