% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassPerm.R
\name{ClassPerm}
\alias{ClassPerm}
\title{simple function to create permutation testing of a classifier}
\usage{
ClassPerm(Data, predictorCol, selectedCols, classifierFun, nSims = 1000, ...)
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

