% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscores.R
\name{fscore}
\alias{fscore}
\title{simple function to generate F-scores (Fisher scores) for ranking features}
\usage{
fscore(Data, featSep, featureCol)
}
\description{
Needs the following:
:
\enumerate{
\item Inputs
  \enumerate{
    \item Data = Data dataframe
     \item featSel = column with different classes
     \item featureCol = all the columns that contain features
}
}
 \enumerate{
\item Outputs
  \enumerate{
    \item named numeric f-scores
}
}
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

