% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscores.R
\name{fscore}
\alias{fscore}
\title{f-score}
\usage{
fscore(Data, featSep, featureCol)
}
\arguments{
\item{Data}{(dataframe) Data dataframe}

\item{featureCol}{(numeric) all the columns that contain features}

\item{featSel}{(numeric) column with different classes}
}
\value{
named numeric \code{f-scores}
}
\description{
A simple function to generate F-scores (Fisher scores) for ranking features
}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

