% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscores.R
\name{fscore}
\alias{fscore}
\title{simple function to generate F-scores (Fisher scores) for ranking features}
\usage{
fscore(Data, featSep, featureCol)
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

