% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDA.R
\name{LinearDA}
\alias{LinearDA}
\title{simple function to perform cross-validated Linear Discriminant Analysis}
\usage{
LinearDA(Data, predictorCol, selectedCols, cvType = "createDataPartition")
}
\description{
Needs the following:
:
\enumerate{
\item Inputs
  \enumerate{
    \item Data = Data dataframe
     \item predictorCol(numeric) = column number that contains the variable to be predicted
     \item selectedCols(numeric) = all the columns of data that would be used either as predictor or as feature
}
}
 \enumerate{
\item Outputs
  \enumerate{
    \item accuracy of discrimination
}
}
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

