% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDA.R
\name{LinearDA}
\alias{LinearDA}
\title{Cross-validated Linear Discriminant Analysis}
\usage{
LinearDA(Data, predictorCol, selectedCols, cvType = "createDataPartition")
}
\arguments{
\item{Data}{(dataframe) Data dataframe}

\item{predictorCol}{(numeric)  column number that contains the variable to be predicted}

\item{selectedCols}{(numeric)  all the columns of data that would be used either as predictor or as feature}
}
\value{
\code{accuracy} of discrimination
}
\description{
A simple function to perform cross-validated Linear Discriminant Analysis
}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

