% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCluster.R
\name{ModelCluster}
\alias{ModelCluster}
\title{Model based Clustering}
\usage{
ModelCluster(Data, NewData = NULL, ...)
}
\arguments{
\item{Data}{(dataframe) Data dataframe}

\item{NewData}{(dataframe) New Data frame for which the class membership is requested}
}
\value{
\code{class membership} of the clustered \code{NewData}
}
\description{
A simple function to perform Model based cluster Analysis
:
}
\author{
Atesh Koul, C'MON group, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

