% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCluster.R
\name{ModelCluster}
\alias{ModelCluster}
\title{simple function to perform Model based cluster Analysis}
\usage{
ModelCluster(Data, NewData = NULL, ...)
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

