% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredPsych-package.r
\docType{package}
\name{PredPsych}
\alias{PredPsych-package}
\alias{PredPsych}
\title{PredPsych.}
\description{
Recent years have seen an increased interest in novel methods for analyzing quantitative data from experimental psychology. Currently, however, they lack an established and accessible software framework. Many existing implementations provide no guidelines, consisting of small code snippets, or sets of packages. In addition, the use of existing packages often requires advanced programming experience. 'PredPsych' is a user-friendly toolbox based on machine learning predictive algorithms. It comprises of multiple functionalities for multivariate analyses of quantitative behavioral data based on machine learning models.
}
\details{
"PredPsych" is a user-friendly, R toolbox based on machine learning predictive algorithms.
}
\references{
Koul, A., Becchio, C., & Cavallo, A. (2017, March 21). 
PredPsych: A toolbox for predictive machine learning based approach in experimental 
psychology research. Retrieved from osf.io/preprints/psyarxiv/pvjac
}
\author{
Atesh Koul, C'MON unit, Istituto Italiano di Tecnologia

\email{atesh.koul@gmail.com}
}
